requirements.txt
streamlit
pandas
google-api-python-client
google-auth-oauthlib